(PC-26667) refactor(apiHelpers): Move ApiError class
