+++ src/features/auth/api/useEmailValidationRemainingResend.ts
@@ -1,7 +1,7 @@
 import { useQuery } from 'react-query'
 
 import { api } from 'api/api'
-import { ApiError } from 'api/apiHelpers'
+import { ApiError } from 'api/ApiError'
 import { QueryKeys } from 'libs/queryKeys'
 
 export const useEmailValidationRemainingResends = ({
