+++ src/features/auth/pages/forgottenPassword/ForgottenPassword/ForgottenPassword.tsx
@@ -4,7 +4,8 @@ import { useForm } from 'react-hook-form'
 import { UseQueryResult } from 'react-query'
 
 import { api } from 'api/api'
-import { ApiError, isAPIExceptionCapturedAsInfo } from 'api/apiHelpers'
+import { ApiError } from 'api/ApiError'
+import { isAPIExceptionCapturedAsInfo } from 'api/apiHelpers'
 import { SettingsResponse } from 'api/gen'
 import { useSettingsContext } from 'features/auth/context/SettingsContext'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
