+++ src/features/bookOffer/pages/BookingOfferModal.tsx
@@ -3,7 +3,8 @@ import React, { useCallback, useEffect, useMemo } from 'react'
 import { useWindowDimensions } from 'react-native'
 import { useTheme } from 'styled-components/native'
 
-import { ApiError, isApiError } from 'api/apiHelpers'
+import { ApiError } from 'api/ApiError'
+import { isApiError } from 'api/apiHelpers'
 import { useBookOfferMutation } from 'features/bookOffer/api/useBookOfferMutation'
 import { BookingCloseInformation } from 'features/bookOffer/components/BookingCloseInformation'
 import { BookingOfferModalFooter } from 'features/bookOffer/components/BookingOfferModalFooter'
