+++ src/features/errors/pages/AsyncErrorBoundary.tsx
@@ -3,7 +3,8 @@ import { FallbackProps } from 'react-error-boundary'
 import { useQueryErrorResetBoundary } from 'react-query'
 import styled from 'styled-components/native'
 
-import { ApiError, isAPIExceptionCapturedAsInfo } from 'api/apiHelpers'
+import { ApiError } from 'api/ApiError'
+import { isAPIExceptionCapturedAsInfo } from 'api/apiHelpers'
 import { homeNavConfig } from 'features/navigation/TabBar/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
 import { AsyncError, MonitoringError, eventMonitoring } from 'libs/monitoring'
