+++ src/features/profile/pages/SuspendAccountConfirmation/SuspendAccountConfirmation.tsx
@@ -4,7 +4,7 @@ import { NativeStackScreenProps } from 'react-native-screens/native-stack'
 import styled from 'styled-components/native'
 
 import { api } from 'api/api'
-import { ApiError } from 'api/apiHelpers'
+import { ApiError } from 'api/ApiError'
 import { navigateToHome } from 'features/navigation/helpers'
 import { RootStackParamList } from 'features/navigation/RootNavigator/types'
 import { useEmailUpdateStatus } from 'features/profile/helpers/useEmailUpdateStatus'
