+++ src/features/trustedDevice/api/useAccountSuspendTokenValidation.ts
@@ -1,7 +1,7 @@
 import { UseQueryResult, useQuery } from 'react-query'
 
 import { api } from 'api/api'
-import { ApiError } from 'api/apiHelpers'
+import { ApiError } from 'api/ApiError'
 import { QueryKeys } from 'libs/queryKeys'
 
 export enum AccountSecurityStatus {
